+++ store/src/main/java/store/domain/goods/controller/model/ImageUrlSet.java
@@ -1,0 +1,18 @@
+package store.domain.goods.controller.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class ImageUrlSet {
+
+    private String imageUrl;
+
+    private String caption;
+
+}
