--- store/src/main/java/store/domain/management/controller/model/ImageUrlSet.java
@@ -1,19 +1,0 @@
-package store.domain.management.controller.model;
-
-import java.util.List;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class ImageUrlSet {
-
-    private String imageUrl;
-
-    private String caption;
-
-}
