--- store/src/main/java/store/domain/management/controller/model/ShippingResponse.java
@@ -1,31 +1,0 @@
-package store.domain.management.controller.model;
-
-import db.domain.receiving.enums.ReceivingStatus;
-import db.domain.shipping.enums.ShippingStatus;
-import jakarta.persistence.Column;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import java.time.LocalDateTime;
-import java.util.List;
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class ShippingResponse {
-
-    private Long id;
-
-    private LocalDateTime deliveryDate;
-
-    private String deliveryAddress;
-
-    private ShippingStatus status;
-
-    private Long userId;
-
-    private Long deliveryMan;
-
-    private List<Long> goodsIdList;
-
-}
