--- store/src/main/java/store/domain/management/converter/GoodsConverter.java
@@ -1,24 +1,0 @@
-package store.domain.management.converter;
-
-import db.domain.goods.GoodsEntity;
-import global.annotation.Converter;
-import store.domain.management.controller.model.GoodsResponse;
-
-@Converter
-public class GoodsConverter {
-
-    public GoodsResponse toResponse(GoodsEntity goodsEntity) {
-        return GoodsResponse.builder()
-            .id(goodsEntity.getId())
-            .name(goodsEntity.getName())
-            .modelName(goodsEntity.getModelName())
-            .category(goodsEntity.getCategory())
-            .quantity(goodsEntity.getQuantity())
-            .abandonmentAt(goodsEntity.getAbandonmentAt())
-            .status(goodsEntity.getStatus())
-            .build();
-    }
-
-
-
-}
