+++ store/src/main/java/store/domain/receiving/converter/ReceivingConverter.java
@@ -1,0 +1,20 @@
+package store.domain.receiving.converter;
+
+import db.domain.receiving.ReceivingEntity;
+import global.annotation.Converter;
+import store.domain.receiving.controller.model.ReceivingResponse;
+
+@Converter
+public class ReceivingConverter {
+
+
+    public ReceivingResponse toResponse(ReceivingEntity receivingEntity) {
+        return ReceivingResponse.builder()
+            .id(receivingEntity.getId())
+            .visitAddress(receivingEntity.getVisitAddress())
+            .visitDate(receivingEntity.getVisitDate())
+            .status(receivingEntity.getStatus())
+            .guaranteeAt(receivingEntity.getGuaranteeAt())
+            .build();
+    }
+}
