+++ store/src/main/resources/templates/addFault.html
@@ -19,9 +19,9 @@
         </a>
   
         <ul class="nav col-12 col-md-auto mb-2 justify-content-center mb-md-0">
-          <li><a href="#" th:href="|@{/api/store/receiving(status=RECEIVING)}|" class="nav-link px-2 link-secondary">입고 관리</a></li>
-          <li><a href="#" th:href="|@{/api/store/shipping(status=PENDING)}|" class="nav-link px-2 link-secondary">출고 관리</a></li>
-          <li><a href="#" th:href="|@{/api/store/goods(status=STORAGE)}|" class="nav-link px-2 link-primary">물품 관리</a></li>
+          <li><a href="#" th:href="|@{/api/receiving(status=RECEIVING)}|" class="nav-link px-2 link-secondary">입고 관리</a></li>
+          <li><a href="#" th:href="|@{/api/shipping(status=PENDING)}|" class="nav-link px-2 link-secondary">출고 관리</a></li>
+          <li><a href="#" th:href="|@{/api/goods(status=STORAGE)}|" class="nav-link px-2 link-primary">물품 관리</a></li>
         </ul>
   
         <div class="col-md-3 text-end">
@@ -39,7 +39,7 @@
         <p class="card-text">카테고리 : <span> 소형 가전 </span></p>
         <p class="card-text">수량 : <span> 1 </span></p>
         
-        <form id="fileForm" action="#" th:action="@{/api/store/fault}" method="post" enctype="multipart/form-data">
+        <form id="fileForm" action="#" th:action="@{/api/fault}" method="post" enctype="multipart/form-data">
           <div id="fileScope">
             <div id="fileElement" class="mb-3">
               <div class="mb-3">
