+++ store/src/main/resources/templates/goodsDetail.html
@@ -24,9 +24,9 @@
       </a>
 
       <ul class="nav col-12 col-md-auto mb-2 justify-content-center mb-md-0">
-        <li><a href="#" th:href="|@{/api/store/receiving(status=RECEIVING)}|" class="nav-link px-2 link-secondary">입고 관리</a></li>
-        <li><a href="#" th:href="|@{/api/store/shipping(status=PENDING)}|" class="nav-link px-2 link-secondary">출고 관리</a></li>
-        <li><a href="#" th:href="|@{/api/store/goods(status=STORAGE)}|" class="nav-link px-2 link-primary">물품 관리</a></li>
+        <li><a href="#" th:href="|@{/api/receiving(status=RECEIVING)}|" class="nav-link px-2 link-secondary">입고 관리</a></li>
+        <li><a href="#" th:href="|@{/api/shipping(status=PENDING)}|" class="nav-link px-2 link-secondary">출고 관리</a></li>
+        <li><a href="#" th:href="|@{/api/goods(status=STORAGE)}|" class="nav-link px-2 link-primary">물품 관리</a></li>
       </ul>
 
       <div class="col-md-3 text-end">
