+++ store/src/main/resources/templates/receivingDetail.html
@@ -24,9 +24,9 @@
       </a>
 
       <ul class="nav col-12 col-md-auto mb-2 justify-content-center mb-md-0">
-        <li><a href="#" th:href="|@{/api/store/receiving(status=RECEIVING)}|" class="nav-link px-2 link-primary">입고 관리</a></li>
-        <li><a href="#" th:href="|@{/api/store/shipping(status=PENDING)}|" class="nav-link px-2 link-secondary">출고 관리</a></li>
-        <li><a href="#" th:href="|@{/api/store/goods(status=STORAGE)}|" class="nav-link px-2 link-secondary">물품 관리</a></li>
+        <li><a href="#" th:href="|@{/api/receiving(status=RECEIVING)}|" class="nav-link px-2 link-primary">입고 관리</a></li>
+        <li><a href="#" th:href="|@{/api/shipping(status=PENDING)}|" class="nav-link px-2 link-secondary">출고 관리</a></li>
+        <li><a href="#" th:href="|@{/api/goods(status=STORAGE)}|" class="nav-link px-2 link-secondary">물품 관리</a></li>
       </ul>
 
       <div class="col-md-3 text-end">
@@ -66,7 +66,7 @@
         </div>
       </div>
     </div>
-    <div class="accordion-item" th:each="goods : ${goodsList}">
+    <div class="accordion-item" th:each="goods : ${receivingRequest.goodsList}">
       <h2 class="accordion-header">
         <button class="accordion-button" th:text="${goods.name}" type="button"
                 data-bs-toggle="collapse" data-bs-target="#collapseOne" aria-expanded="true"
@@ -135,7 +135,7 @@
               <div th:if="${receivingRequest.status.name() == 'LOADING'}" class="row">
                 <div class="col"></div>
                 <div class="col-8">
-                  <a type="button" th:href="@{|/api/store/fault/${goods.id}|}" class="btn btn-danger mt-3 w-100">결함 등록</a>
+                  <a type="button" th:href="@{|/api/fault/${goods.id}|}" class="btn btn-danger mt-3 w-100">결함 등록</a>
                 </div>
                 <div class="col"></div>
               </div>
@@ -150,10 +150,10 @@
 
   <!-- arcod end -->
 
-  <form th:if="${receivingRequest.status.name() == 'RECEIVING'}" action="" th:action="@{|/api/store/loading/${receivingRequest.id}|}" method="post">
+  <form th:if="${receivingRequest.status.name() == 'RECEIVING'}" action="" th:action="@{|/api/loading/${receivingRequest.id}|}" method="post">
     <button type="summit" class="btn btn-primary w-100 mb-3 mt-3">입고 시작</button>
   </form>
-  <a th:if="${receivingRequest.status.name() == 'LOADING'}" type="button" th:href="@{|/api/store/storage/${receivingRequest.id}|}"  class="btn btn-primary w-100 mb-3 mt-3">창고 배치</a>
+  <a th:if="${receivingRequest.status.name() == 'LOADING'}" type="button" th:href="@{|/api/store/${receivingRequest.id}|}"  class="btn btn-primary w-100 mb-3 mt-3">창고 배치</a>
 
 </div>
