+++ store/src/main/resources/templates/shippingDetail.html
@@ -24,9 +24,9 @@
       </a>
 
       <ul class="nav col-12 col-md-auto mb-2 justify-content-center mb-md-0">
-        <li><a href="#" th:href="|@{/api/store/receiving(status=RECEIVING)}|" class="nav-link px-2 link-secondary">입고 관리</a></li>
-        <li><a href="#" th:href="|@{/api/store/shipping(status=PENDING)}|" class="nav-link px-2 link-primary">출고 관리</a></li>
-        <li><a href="#" th:href="|@{/api/store/goods(status=STORAGE)}|" class="nav-link px-2 link-secondary">물품 관리</a></li>
+        <li><a href="#" th:href="|@{/api/receiving(status=RECEIVING)}|" class="nav-link px-2 link-secondary">입고 관리</a></li>
+        <li><a href="#" th:href="|@{/api/shipping(status=PENDING)}|" class="nav-link px-2 link-primary">출고 관리</a></li>
+        <li><a href="#" th:href="|@{/api/goods(status=STORAGE)}|" class="nav-link px-2 link-secondary">물품 관리</a></li>
       </ul>
 
       <div class="col-md-3 text-end">
@@ -66,7 +66,7 @@
         </div>
       </div>
     </div>
-    <div class="accordion-item" th:each="goods : ${goodsList}">
+    <div class="accordion-item" th:each="goods : ${shippingRequest.goodsList}">
       <h2 class="accordion-header">
         <button class="accordion-button" th:text="${goods.name}" type="button"
                 data-bs-toggle="collapse" data-bs-target="#collapseOne" aria-expanded="true"
